// frontend/src/api/generated/types.gen.ts
export type ProfileWrite = {
    experienceLevel: 'beginner' | 'intermediate' | 'advanced' | 'expert';
    displayUnits: 'metric' | 'imperial';
};

export type ProfileUpdate = {
    experienceLevel: 'beginner' | 'intermediate' | 'advanced' | 'expert';
    displayUnits: 'metric' | 'imperial';
    expectedVersion: number;
};

export type Profile = {
    experienceLevel: 'beginner' | 'intermediate' | 'advanced' | 'expert';
    displayUnits: 'metric' | 'imperial';
    version: number;
    createdAt: string;
    updatedAt: string;
};

export type SpotWrite = {
    name: string;
    longitude: number;
    latitude: number;
    /**
     * IANA time zone identifier
     */
    timeZone: string;
};

export type SpotUpdate = {
    name: string;
    longitude: number;
    latitude: number;
    timeZone: string;
    expectedVersion: number;
};

export type Spot = {
    id: string;
    name: string;
    longitude: number;
    latitude: number;
    timeZone: string;
    version: number;
    createdAt: string;
    updatedAt: string;
};

export type SpotList = {
    items: Array<Spot>;
};

export type FavoriteWrite = {
    spotId: string;
    sortPosition: number;
};

export type FavoritePosition = {
    sortPosition: number;
};

export type Favorite = {
    spotId: string;
    sortPosition: number;
    createdAt: string;
    updatedAt: string;
};